Fixed bugs related to averaging scalar fields.
Minimum normal estimation is now based on K nearest neighbors to avoid
NaN values.

Separated getNormals into getNormalsRadiusSearch and KSearch.

Updated methodws throughout.

+++ src/ScalarField.cpp
@@ -12,7 +12,7 @@ namespace GeoDetection
 		const float range = *max - *min;
 
 #pragma omp parallel for
-		for (size_t i = 0; i < data.size(); i++)
+		for (int64_t i = 0; i < (int64_t)data.size(); i++)
 		{
 			data[i] = (data[i] - *min) / range;
 		}
